+++ lib/questionnaires/view/item/answer/src/coding_answer_filler.dart
@@ -275,7 +275,7 @@ Widget _createStyledOption(
   final optionPrefix = optionModel.optionPrefix;
   final optionText = optionModel.optionText;
 
-  final optionTitle = <XhtmlString>[
+  final optionTitle = <RenderingString>[
     if (optionPrefix != null) optionPrefix,
     optionText,
   ].concatenateXhtml(': ');